WIP #89: Update the script that generates the training samples.

+++ studies/ntuple-RDX_l0_hadron_tos_training_sample/gen_l0hadron_samples.py
@@ -3,20 +3,24 @@
 # Script to generate various samples for the trigger emulation
 
 import os
-import pathlib
 import sys
 
-# ntpIn = '../../ntuples/0.9.4-trigger_emulation/Dst_D0-mc/Dst_D0--21_04_21--mc--MC_2016_Beam6500GeV-2016-MagDown-Nu1.6-25ns-Pythia8_Sim09j_Trig0x6139160F_Reco16_Turbo03a_Filtered_11574021_D0TAUNU.SAFESTRIPTRIG.DST.root'
-ntpIn = '../../ntuples/0.9.5-bugfix/Dst_D0-mc/Dst_D0--21_10_08--mc--MC_2016_Beam6500GeV-2016-MagDown-Nu1.6-25ns-Pythia8_Sim09j_Trig0x6139160F_Reco16_Turbo03a_Filtered_11574021_D0TAUNU.SAFESTRIPTRIG.DST.root'
+from os.path import isfile
+from os.path import splitext
 
-if not pathlib.Path(ntpIn).is_file():
-    sys.exit(ntpIn+' does not exist, you need to download it')
 
+###########
+# Helpers #
+###########
 
 def runCmd(cmd):
     print('  \033[92m'+cmd+'\033[0m')
     os.system(cmd)
 
+
+######################
+# Generating ntuples #
+######################
 
 def slim(tag, ntpIn):
     yml = 'l0hadron_sample_'+tag+'.yml'
@@ -25,19 +29,45 @@ def slim(tag, ntpIn):
     return ntpOut
 
 
+def merge(ntpOut, ntpsIn):
+    runCmd('hadd -fk {} {}'.format(ntpOut, ' '.join(ntpsIn)))
+    return ntpOut
+
+
+def rename(ntpOut, ntpIn):
+    runCmd('cp {} {}'.format(ntpIn, ntpOut))
+    return ntpOut
+
+
+def splitTrainValid(ntpIn):
+    base = splitext(ntpIn)[0]
+    ntpsOut = [base+'_'+mode+'.root' for mode in ['train', 'valid', 'test']]
+    runCmd('root -l \'../../scripts/split_train_vali_test.C("'+ntpIn+'", "50:50")\'')
+    return ntpsOut
+
+
+# ntpIn = '../../ntuples/0.9.4-trigger_emulation/Dst_D0-mc/Dst_D0--21_04_21--mc--MC_2016_Beam6500GeV-2016-MagDown-Nu1.6-25ns-Pythia8_Sim09j_Trig0x6139160F_Reco16_Turbo03a_Filtered_11574021_D0TAUNU.SAFESTRIPTRIG.DST.root'
+ntpIn = '../../ntuples/0.9.5-bugfix/Dst_D0-mc/Dst_D0--21_10_08--mc--MC_2016_Beam6500GeV-2016-MagDown-Nu1.6-25ns-Pythia8_Sim09j_Trig0x6139160F_Reco16_Turbo03a_Filtered_11574021_D0TAUNU.SAFESTRIPTRIG.DST.root'
+
+if not isfile(ntpIn):
+    sys.exit(ntpIn+' does not exist, you need to download it')
+
 ## Slimming the trigger-matched ntuple
-ntpNtm = slim('tm', ntpIn)
-runCmd('root -l \'../../scripts/split_train_vali_test.C("'+ntpNtm+'", "50:50")\'')
+ntpTm = slim('tm', ntpIn)
+ntpTmTrain, ntpTmValid, _ = splitTrainValid(ntpTm)
 
 ## Slimming the non-trigger-matched ntuple
 ntpNtm = slim('ntm', ntpIn)
-runCmd('root -l \'../../scripts/split_train_vali_test.C("'+ntpNtm+'", "50:50")\'')
+ntpNtmTrain, ntpNtmValid, _ = splitTrainValid(ntpNtm)
 
 ## Merge the training samples for XGB
-runCmd('hadd -fk run2-rdx-train_xgb.root l0hadron_emu_tm_train.root l0hadron_emu_ntm_train.root')
+ntpTrainXgb = merge('run2-rdx-train_xgb.root', [ntpTmTrain, ntpNtmTrain])
+
+## Merge the validation samples
+ntpValid = merge('run2-rdx-valid.root', [ntpTmValid, ntpNtmValid])
 
 ## Only use trigger-matched training sample for BDG
-runCmd('mv ./l0hadron_emu_tm_train.root ./run2-rdx-train_bdt.root')
+ntpTrainBdt = rename('run2-rdx-train_bdt.root', ntpTmTrain)
 
 ## Remove unused ntuples
 runCmd('rm l0hadron_emu*.root')
